requirements.txt
playwright>=1.40.0
python-telegram-bot>=20.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Testing dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0
